python-application/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
webrtcvad-wheels==2.0.14
numpy==1.24.3
scipy==1.11.4
torch==2.1.1
torchaudio==2.1.1
transformers==4.35.2
tensorflow==2.15.0
tensorflow-hub==0.15.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
psycopg2-binary==2.9.9
Pillow==10.1.0
requests==2.31.0
